refactor(telegrambot): put the ten stragglers in the project's own namespace
Ten files in TallaEgg.TelegramBot.Infrastructure declared `namespace
TallaEgg.TelegramBot` -- the name of a project that was deleted -- while the
other twenty-two in the same project declared TallaEgg.TelegramBot.Infrastructure.
It compiled because C# walks outward from the enclosing namespace, so a type in
the parent was visible without a using, which is exactly why the split could
persist unnoticed.

The ten move to TallaEgg.TelegramBot.Infrastructure, and the two under Clients/
to TallaEgg.TelegramBot.Infrastructure.Clients, matching OrderApiClient and
IOrderApiClient beside them. Twelve consumers carried `using
TallaEgg.TelegramBot;`: nine already had the Infrastructure using so the line is
simply gone, two now name Infrastructure instead, and TelegramBotHostedService
needed neither since it is in that namespace itself.

One line per file, thirty-four in all, no other content. Behaviour is unchanged:
nothing resolves these types by name, and no configuration, project file or
workflow mentions the namespace.

IAffiliateApiClient.cs nearly escaped the sweep -- its BOM puts `namespace` on
line 1, so an anchored grep skips it and only the compiler noticed.
