+++ TelegramBot/TallaEgg.TelegramBot.Infrastructure/ProxyTest.cs
@@ -3,7 +3,7 @@ using System.Net;
 using System.Net.Http;
 using System.Threading.Tasks;
 
-namespace TallaEgg.TelegramBot
+namespace TallaEgg.TelegramBot.Infrastructure
 {
     public class ProxyTest
     {
