+++ TelegramBot/TallaEgg.TelegramBot.Simulator/Program.cs
@@ -4,7 +4,6 @@ using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using TallaEgg.Core.Services;
 using TallaEgg.Infrastructure.Clients;
-using TallaEgg.TelegramBot;
 using TallaEgg.TelegramBot.Core.Interfaces;
 using TallaEgg.TelegramBot.Infrastructure;
 using TallaEgg.TelegramBot.Infrastructure.Clients;
