+++ tests/TallaEgg.AllServices.Tests/AdminChargeCommandTests.cs
@@ -6,7 +6,6 @@ using TallaEgg.Core.DTOs.User;
 using TallaEgg.Core.DTOs.Wallet;
 using TallaEgg.Core.Enums.User;
 using TallaEgg.Core.Requests.Wallet;
-using TallaEgg.TelegramBot;
 using TallaEgg.TelegramBot.Infrastructure;
 using TallaEgg.TelegramBot.Infrastructure.Conversations;
 using Telegram.Bot.Types;
