+++ tests/TallaEgg.AllServices.Tests/AutoQuoteCommandTests.cs
@@ -2,7 +2,6 @@
 using TallaEgg.Core;
 using TallaEgg.Core.DTOs.User;
 using TallaEgg.Core.Enums.User;
-using TallaEgg.TelegramBot;
 using TallaEgg.TelegramBot.Infrastructure;
 using TallaEgg.TelegramBot.Infrastructure.Conversations;
 using Telegram.Bot.Types;
