+++ tests/TallaEgg.AllServices.Tests/BotConversationFlowTests.cs
@@ -4,7 +4,6 @@ using TallaEgg.Core.DTOs.Order;
 using TallaEgg.Core.DTOs.User;
 using TallaEgg.Core.Enums.Order;
 using TallaEgg.Core.Enums.User;
-using TallaEgg.TelegramBot;
 using TallaEgg.TelegramBot.Infrastructure;
 using TallaEgg.TelegramBot.Infrastructure.Conversations;
 using TallaEgg.AllServices.Tests.Fakes;
