+++ tests/TallaEgg.AllServices.Tests/RegistrationMessagingTests.cs
@@ -1,7 +1,6 @@
 using Microsoft.Extensions.Logging.Abstractions;
 using TallaEgg.Core.DTOs.User;
 using TallaEgg.Core.Enums.User;
-using TallaEgg.TelegramBot;
 using TallaEgg.TelegramBot.Infrastructure;
 using TallaEgg.TelegramBot.Infrastructure.Conversations;
 using TallaEgg.TelegramBot.Infrastructure.Extensions.Telegram;
